reorganized components and installed material-ui
