+++ client/src/components/journeyList.tsx
@@ -1,0 +1,7 @@
+const JourneyList = () => {
+    return (
+        <div>journey list</div>
+    )
+}
+
+export default JourneyList;
